update jwt-scala-hardcode rule
