--- scala/jwt-scala/security/jwt-play-json-hardcode.yaml
@@ -1,30 +1,0 @@
-rules:
-- id: jwt-play-json-hardcode
-  patterns:
-    - pattern-inside: |
-        import pdi.jwt.$DEPS
-        ...
-    - pattern-either:
-      - patterns: |
-          JwtJson.$METHOD($CLAIM, "...", ...)
-      - patterns:
-        - pattern: JwtJson.$METHOD($CLAIM, this.$KEY, ...)
-        - pattern-inside: |
-            class $CLASS {
-              ...
-              $KEY = "..."
-              ...
-            }
-  message: >-
-    Hardcoded JWT secret or private key is used.
-    This is a Insufficiently Protected Credentials weakness: https://cwe.mitre.org/data/definitions/522.html
-    Consider using an appropriate security mechanism to protect the credentials (e.g. keeping secrets in environment variables)
-  languages: [scala]
-  severity: WARNING
-  metadata: 
-      category: security
-      cwe: "CWE-78: Improper Neutralization of Special Elements used in an OS Command
-        ('OS Command Injection')"
-      owasp: "A1: Injection"
-      technology:
-        - scala
